Code Review 이후 코드 중복 및 개선
- ParcelInformation 파일 분리
- InvoiceView에서 고민했던 4,9원칙 적용 !
